+++ output_example_input.txt
@@ -1,12 +1,12 @@
 >Sequence one
-GATGCTAGCTGATCGTGCGCGCGCGCTAGCTGATCGTAGTCGATGCTGCGCGCGGCTGTAGCTGATCGATGCTACGCGCGATGCTAGTCGAGTCGATGCTAGCTGATCG
+TAGAGCTCGCGATGCTAGCTGATCGTGCGCGCGCGCTAGCTGATCGTAGTCGATGCTGCGCGCGGCTGTAGCTGATCGATGCTACGCGCGATGCTAGTCGAGTCGATGC
 
 >Sequence two
-GATGCTAGCTGATCGTGCGCGCGCGCTAGCTGATCGTAGTCGATGCTGCGCGCGGCTGTAGCTGATCGATGCTACGCGCGATGCTAGTCGAGTCGATGCTAGCTGATCGATGCTAGCTGGAGAGATCGCTAGAGCTCGCGATGCTAGCTGATCGTGCGCGCGCGCTAGCTGATCGTAGTCGATGCTGCGCGCGGCTGTAGCTGATCGATGCTACGCGCGATGCTAGTCGAGTCGATGCTAGCTGATCGATGCTAGCTGGAGAGATCGCTAGAGCTCGCGATGCTAGCTGATCGTGCGCGCGCGCTAGCTGATCGTAGTCGATGCTGCGCGCGGCTGTAGCTGATCGATGCTACGCGCGATGCTAGTCGAGTCGATGCTAGCTGATCG
+TAGAGCTCGCGATGCTAGCTGATCGTGCGCGCGCGCTAGCTGATCGTAGTCGATGCTGCGCGCGGCTGTAGCTGATCGATGCTACGCGCGATGCTAGTCGAGTCGATGCTAGCTGATCGATGCTAGCTGGAGAGATCGCTAGAGCTCGCGATGCTAGCTGATCGTGCGCGCGCGCTAGCTGATCGTAGTCGATGCTGCGCGCGGCTGTAGCTGATCGATGCTACGCGCGATGCTAGTCGAGTCGATGCTAGCTGATCGATGCTAGCTGGAGAGATCGCTAGAGCTCGCGATGCTAGCTGATCGTGCGCGCGCGCTAGCTGATCGTAGTCGATGCTGCGCGCGGCTGTAGCTGATCGATGCTACGCGCGATGCTAGTCGAGTCGATGC
 
 >Sequence three
-GATGCTAGCTGATCGTGCGCGCGCGCTAGCTGATCGTAGTCGATGCTGCGCGCGGCTGTAGCTGATCGATGCTACGCGCGATGCTAGTCGAGTCGATGCTAGCTGATCGATGCTAGCTGGAGAGATCGCTAGAGCTCGCGATGCTAGCTGATCGTGCGCGCGCGCTAGCTGATCGTAGTCGATGCTGCGCGCGGCTGTAGCTGATCGATGCTACGCGCGATGCTAGTCGAGTCGATGCTAGCTGATCGATGCTAGCTGGAGAGATCGCTAGAGCTCGCGATGCTAGCTGATCGTGCGCGCGCGCTAGCTGATCGTAGTCGATGCTGCGCGCGGCTGTAGCTGATCGATGCTACGCGCGATGCTAGTCGAGTCGATGCTAGCTGATCGATGCTAGCTGGAGAGATCGCTAGAGCTCGCGATGCTAGCTGATCGTGCGCGCGCGCTAGCTGATCGTAGTCGATGCTGCGCGCGGCTGTAGCTGATCGATGCTACGCGCGATGCTAGTCGAGTCGATGCTAGCTGATCGATGCTAGCTGGAGAGATCGCTAGAGCTCGCGATGCTAGCTGATCGTGCGCGCGCGCTAGCTGATCGTAGTCGATGCTGCGCGCGGCTGTAGCTGATCGATGCTACGCGCGATGCTAGTCGAGTCGATGCTAGCTGATCGATGCTAGCTGGAGAGATCGCTAGAGCTCGCGATGCTAGCTGATCGTGCGCGCGCGCTAGCTGATCGTAGTCGATGCTGCGCGCGGCTGTAGCTGATCGATGCTACGCGCGATGCTAGTCGAGTCGATGCTAGCTGATCGATGCTAGCTGGAGAGATCGCTAGAGCTCGCGATGCTAGCTGATCGTGCGCGCGCGCTAGCTGATCGTAGTCGATGCTGCGCGCGGCTGTAGCTGATCGATGCTACGCGCGATGCTAGTCGAGTCGATGCTAGCTGATCGATGCTAGCTGGAGAGATCGCTAGAGCTCGCGATGCTAGCTGATCGTGCGCGCGCGCTAGCTGATCGTAGTCGATGCTGCGCGCGGCTGTAGCTGATCGATGCTACGCGCGATGCTAGTCGAGTCGATGCTAGCTGATCG
+TAGAGCTCGCGATGCTAGCTGATCGTGCGCGCGCGCTAGCTGATCGTAGTCGATGCTGCGCGCGGCTGTAGCTGATCGATGCTACGCGCGATGCTAGTCGAGTCGATGCTAGCTGATCGATGCTAGCTGGAGAGATCGCTAGAGCTCGCGATGCTAGCTGATCGTGCGCGCGCGCTAGCTGATCGTAGTCGATGCTGCGCGCGGCTGTAGCTGATCGATGCTACGCGCGATGCTAGTCGAGTCGATGCTAGCTGATCGATGCTAGCTGGAGAGATCGCTAGAGCTCGCGATGCTAGCTGATCGTGCGCGCGCGCTAGCTGATCGTAGTCGATGCTGCGCGCGGCTGTAGCTGATCGATGCTACGCGCGATGCTAGTCGAGTCGATGCTAGCTGATCGATGCTAGCTGGAGAGATCGCTAGAGCTCGCGATGCTAGCTGATCGTGCGCGCGCGCTAGCTGATCGTAGTCGATGCTGCGCGCGGCTGTAGCTGATCGATGCTACGCGCGATGCTAGTCGAGTCGATGCTAGCTGATCGATGCTAGCTGGAGAGATCGCTAGAGCTCGCGATGCTAGCTGATCGTGCGCGCGCGCTAGCTGATCGTAGTCGATGCTGCGCGCGGCTGTAGCTGATCGATGCTACGCGCGATGCTAGTCGAGTCGATGCTAGCTGATCGATGCTAGCTGGAGAGATCGCTAGAGCTCGCGATGCTAGCTGATCGTGCGCGCGCGCTAGCTGATCGTAGTCGATGCTGCGCGCGGCTGTAGCTGATCGATGCTACGCGCGATGCTAGTCGAGTCGATGCTAGCTGATCGATGCTAGCTGGAGAGATCGCTAGAGCTCGCGATGCTAGCTGATCGTGCGCGCGCGCTAGCTGATCGTAGTCGATGCTGCGCGCGGCTGTAGCTGATCGATGCTACGCGCGATGCTAGTCGAGTCGATGCTAGCTGATCGATGCTAGCTGGAGAGATCGCTAGAGCTCGCGATGCTAGCTGATCGTGCGCGCGCGCTAGCTGATCGTAGTCGATGCTGCGCGCGGCTGTAGCTGATCGATGCTACGCGCGATGCTAGTCGAGTCGATGC
 
 >Sequence FOUR
-GATGCTAGCTGATCGTGCGCGCGCGCTAGCTGATCGTAGTCGATGCTGCGCGCGGCTGTAGCTGATCGATGCTACGCGCGATGCTAGTCGAGTCGATGCTAGCTGATCGATGCTAGCTGGAGAGATCGCTAGAGCTCGCGATGCTAGCTGATCGTGCGCGCGCGCTAGCTGATCGTAGTCGATGCTGCGCGCGGCTGTAGCTGATCGATGCTACGCGCGATGCTAGTCGAGTCGATGCTAGCTGATCGATGCTAGCTGGAGAGATCGCTAGAGCTCGCGATGCTAGCTGATCGTGCGCGCGCGCTAGCTGATCGTAGTCGATGCTGCGCGCGGCTGTAGCTGATCGATGCTACGCGCGATGCTAGTCGAGTCGATGCTAGCTGATCGATGCTAGCTGGAGAGATCGCTAGAGCTCGCGATGCTAGCTGATCGTGCGCGCGCGCTAGCTGATCGTAGTCGATGCTGCGCGCGGCTGTAGCTGATCGATGCTACGCGCGATGCTAGTCGAGTCGATGCTAGCTGATCGATGCTAGCTGGAGAGATCGCTAGAGCTCGCGATGCTAGCTGATCGTGCGCGCGCGCTAGCTGATCGTAGTCGATGCTGCGCGCGGCTGTAGCTGATCGATGCTACGCGCGATGCTAGTCGAGTCGATGCTAGCTGATCGATGCTAGCTGGAGAGATCGCTAGAGCTCGCGATGCTAGCTGATCGTGCGCGCGCGCTAGCTGATCGTAGTCGATGCTGCGCGCGGCTGTAGCTGATCGATGCTACGCGCGATGCTAGTCGAGTCGATGCTAGCTGATCGATGCTAGCTGGAGAGATCGCTAGAGCTCGCGATGCTAGCTGATCGTGCGCGCGCGCTAGCTGATCGTAGTCGATGCTGCGCGCGGCTGTAGCTGATCGATGCTACGCGCGATGCTAGTCGAGTCGATGCTAGCTGATCGATGCTAGCTGGAGAGATCGCTAGAGCTCGCGATGCTAGCTGATCGTGCGCGCGCGCTAGCTGATCGTAGTCGATGCTGCGCGCGGCTGTAGCTGATCGATGCTACGCGCGATGCTAGTCGAGTCGATGCTAGCTGATCGATGCTAGCTGGAGAGATCGCTAGAGCTCGCGATGCTAGCTGATCGTGCGCGCGCGCTAGCTGATCGTAGTCGATGCTGCGCGCGGCTGTAGCTGATCGATGCTACGCGCGATGCTAGTCGAGTCGATGCTAGCTGATCGATGCTAGCTGGAGAGATCGCTAGAGCTCGCGATGCTAGCTGATCGTGCGCGCGCGCTAGCTGATCGTAGTCGATGCTGCGCGCGGCTGTAGCTGATCGATGCTACGCGCGATGCTAGTCGAGTCGATGCTAGCTGATCG
+TAGAGCTCGCGATGCTAGCTGATCGTGCGCGCGCGCTAGCTGATCGTAGTCGATGCTGCGCGCGGCTGTAGCTGATCGATGCTACGCGCGATGCTAGTCGAGTCGATGCTAGCTGATCGATGCTAGCTGGAGAGATCGCTAGAGCTCGCGATGCTAGCTGATCGTGCGCGCGCGCTAGCTGATCGTAGTCGATGCTGCGCGCGGCTGTAGCTGATCGATGCTACGCGCGATGCTAGTCGAGTCGATGCTAGCTGATCGATGCTAGCTGGAGAGATCGCTAGAGCTCGCGATGCTAGCTGATCGTGCGCGCGCGCTAGCTGATCGTAGTCGATGCTGCGCGCGGCTGTAGCTGATCGATGCTACGCGCGATGCTAGTCGAGTCGATGCTAGCTGATCGATGCTAGCTGGAGAGATCGCTAGAGCTCGCGATGCTAGCTGATCGTGCGCGCGCGCTAGCTGATCGTAGTCGATGCTGCGCGCGGCTGTAGCTGATCGATGCTACGCGCGATGCTAGTCGAGTCGATGCTAGCTGATCGATGCTAGCTGGAGAGATCGCTAGAGCTCGCGATGCTAGCTGATCGTGCGCGCGCGCTAGCTGATCGTAGTCGATGCTGCGCGCGGCTGTAGCTGATCGATGCTACGCGCGATGCTAGTCGAGTCGATGCTAGCTGATCGATGCTAGCTGGAGAGATCGCTAGAGCTCGCGATGCTAGCTGATCGTGCGCGCGCGCTAGCTGATCGTAGTCGATGCTGCGCGCGGCTGTAGCTGATCGATGCTACGCGCGATGCTAGTCGAGTCGATGCTAGCTGATCGATGCTAGCTGGAGAGATCGCTAGAGCTCGCGATGCTAGCTGATCGTGCGCGCGCGCTAGCTGATCGTAGTCGATGCTGCGCGCGGCTGTAGCTGATCGATGCTACGCGCGATGCTAGTCGAGTCGATGCTAGCTGATCGATGCTAGCTGGAGAGATCGCTAGAGCTCGCGATGCTAGCTGATCGTGCGCGCGCGCTAGCTGATCGTAGTCGATGCTGCGCGCGGCTGTAGCTGATCGATGCTACGCGCGATGCTAGTCGAGTCGATGCTAGCTGATCGATGCTAGCTGGAGAGATCGCTAGAGCTCGCGATGCTAGCTGATCGTGCGCGCGCGCTAGCTGATCGTAGTCGATGCTGCGCGCGGCTGTAGCTGATCGATGCTACGCGCGATGCTAGTCGAGTCGATGCTAGCTGATCGATGCTAGCTGGAGAGATCGCTAGAGCTCGCGATGCTAGCTGATCGTGCGCGCGCGCTAGCTGATCGTAGTCGATGCTGCGCGCGGCTGTAGCTGATCGATGCTACGCGCGATGCTAGTCGAGTCGATGC
